refactor(schema): move http engine to its own file
